requirements.txt
gradio>=4.31.0
pandas
matplotlib
scikit-learn
holidays
statsmodels
numpy>=1.24
plotly
dash
bcrypt
pyotp
qrcode
pillow

